requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
attrs==25.4.0
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cryptography==46.0.3
discord.py==2.6.4
fastapi==0.116.1
firebase_admin==7.1.0
frozenlist==1.8.0
google-api-core==2.27.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
msgpack==1.1.2
multidict==6.7.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.22.0
